+++ CHANGELOG.md
@@ -21,8 +21,9 @@
 * A bunch of new advancements.
 * A bunch of new Shotlocks:
   * Volley style: Meteor Shower, Flame Salvo, Chaos Snake. 
-  * Dashing style: Absolute Zero, Photon Charge, Lightning Ray. 
+  * Dashing style: Sonic Blade, Absolute Zero, Photon Charge, Lightning Ray. 
   * Circular style: Bubble Blaster, Bio Barrage, Thunderstorm, Pulse Bomb.
+  * Special Style: Ultima Cannon.
 * Toggle Gummi Ship flight mode with pick block key (by default middle mouse click).
   * In that flight mode the ship follows the camera, similar to elytra.
 * Recipe to craft the dev heads.
@@ -43,7 +44,7 @@
 * Slightly optimized some parts of the code.
 * Updated Keyblade weapon to match new EFM declarations.
 * White Mushroom rewards are now data driven, by default the loot table is composed by orichalcum, orichalcum plus, illusory crystal, evanescent crystal, manifest illusion and lost illusion, and they drop from 1 up to 2 or 3 of one material.
-* Datagen'd Org limits.
+* Datagen'd Org limits and levelling paths.
 
 ## Fixed:
 * Castle Oblivion door criteria values match COM's mechanics, the doors within a generated room will be +1 the value of the card until it reaches 9 then will be 0.
@@ -54,4 +55,4 @@
 * Being able to use air dash while in a vehicle.
 * Removed ship fall damage for non player entities.
 * Fixed Gummi Hangar hologram not rendering.
-* Orientation of thrown lances and cards.
+* Orientation of thrown lances, scythes and cards.